% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cache.info}
\alias{cache.info}
\title{Returns information about the cache.}
\usage{
cache.info(cache.dir = "cache", units = "mins", stale = c(hourly = hourly,
  daily = daily, weekly = weekly, monthly = monthly))
}
\arguments{
\item{cache.dir}{the directory containing the cached files.}

\item{units}{the units to use for calculate the age of the cache file.}

\item{stale}{a vector of frequencies to test whether the each cache file
is stale.}
}
\value{
a data frame with three columns: the cached file name, the date/time
        created, and the age in the specified units (default is minutes).
}
\description{
Returns information about the cache.
}

