% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cacheData}
\alias{cacheData}
\title{Retrieve data from a data cache.}
\usage{
cacheData(FUN, frequency = hourly, cache.dir = "cache",
  envir = parent.frame(), ...)
}
\arguments{
\item{FUN}{the function used to laod the data.}

\item{cache.dir}{the directory containing the cached data files.}

\item{envir}{the enviroment into which data will be loaded.}

\item{...}{other parameters passed to \code{FUN}.}
}
\description{
Retrieve data from a data cache.
}

