% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data.cache}
\alias{data.cache}
\title{Retrieve data from a data cache.}
\usage{
data.cache(FUN, frequency = daily, cache.dir = "cache",
  cache.name = "Cache-", envir = parent.frame(), wait = FALSE, ...)
}
\arguments{
\item{FUN}{the function used to laod the data.}

\item{frequency}{how frequently should the cache expire.}

\item{cache.dir}{the directory containing the cached data files.}

\item{envir}{the enviroment into which data will be loaded.}

\item{wait}{should the function wait until stale data is refreshed.}

\item{...}{other parameters passed to \code{FUN}.}
}
\description{
Retrieve data from a data cache.
}
\seealso{
\link{daily}, \link{hourly}, \link{weekly}, \link{monthly},
       \link{yearly}, \link{nMinutes}, \link{nHours}, \link{nDays}
}

