% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hourly}
\alias{hourly}
\title{Refresh data hourly.}
\usage{
hourly(timestamp)
}
\arguments{
\item{timestamp}{the timestamp to test whether the cache is stale.}
}
\value{
Returns TRUE if the the cache is stale.
}
\description{
This function will return TRUE when the data cache is stale and the data should
be refreshed. Essentially this will return TRUE after the top of each hour.
For example, if the last cache was created at 9:59 and then called again at
10:01, this function will return TRUE. If you wish to refresh every, say, 60
minutes, use the \code{\link{nMinutes}} function.
}
\seealso{
Other frequencies: \code{\link{daily}};
  \code{\link{monthly}}; \code{\link{nDays}};
  \code{\link{nHours}}; \code{\link{nMinutes}};
  \code{\link{weekly}}; \code{\link{yearly}}
}

