% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies.R
\name{monthly}
\alias{monthly}
\title{Refresh data monthly.}
\usage{
monthly(timestamp)
}
\arguments{
\item{timestamp}{the timestamp to test whether the cache is stale.}
}
\description{
Data becomes stale monthly.
}
\seealso{
Other frequencies: \code{\link{daily}},
  \code{\link{hourly}}, \code{\link{nDays}},
  \code{\link{nHours}}, \code{\link{nMinutes}},
  \code{\link{now}}, \code{\link{weekly}},
  \code{\link{yearly}}
}

