% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nDays}
\alias{nDays}
\title{Refresh every n days.}
\usage{
nDays(nDays)
}
\arguments{
\item{nDays}{number of days (minimally) between updates.}
}
\description{
Refresh every n days.
}
\seealso{
Other frequencies: \code{\link{daily}};
  \code{\link{hourly}}; \code{\link{monthly}};
  \code{\link{nHours}}; \code{\link{nMinutes}};
  \code{\link{weekly}}; \code{\link{yearly}}
}

