% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies.R
\name{nMinutes}
\alias{nMinutes}
\title{Refresh every n minutes.}
\usage{
nMinutes(minutes)
}
\arguments{
\item{minutes}{number of minutes (minimally) between updates.}
}
\description{
Data becomes stale after n minutes.
}
