% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies.R
\name{now}
\alias{now}
\title{Refresh the cache now.}
\usage{
now(timestamp)
}
\arguments{
\item{timestamp}{the timestamp to test whether the cache is stale.}
}
\value{
Always returns TRUE.
}
\description{
This function will always return TRUE therefore causing the cache to always
be refreshed.
}
