% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies.R
\name{yearly}
\alias{yearly}
\title{Refresh data yearly.}
\usage{
yearly(timestamp)
}
\arguments{
\item{timestamp}{the timestamp to test whether the cache is stale.}
}
\description{
Refresh data yearly.
}
\seealso{
Other frequencies: \code{\link{daily}},
  \code{\link{hourly}}, \code{\link{monthly}},
  \code{\link{nDays}}, \code{\link{nHours}},
  \code{\link{nMinutes}}, \code{\link{now}},
  \code{\link{weekly}}
}

