% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cache.info}
\alias{cache.info}
\title{Returns information about the cache.}
\usage{
cache.info(cache.dir = "cache", cache.name = "Cache-", units = "mins",
  stale = c(hourly = hourly, daily = daily, weekly = weekly, monthly =
  monthly, yearly = yearly))
}
\arguments{
\item{cache.dir}{the directory containing the cached files.}

\item{units}{the units to use for calculate the age of the cache file.}

\item{stale}{a vector of frequencies to test whether each cache file
is stale according to that metric. If \code{NULL}, no info is provided.}
}
\value{
a data frame with three columns: the cached file name, the date/time
        created, and the age in the specified units (default is minutes).
}
\description{
Returns information about the cache.
}

