% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data.cache}
\alias{data.cache}
\title{Retrieve data from a data cache.}
\usage{
data.cache(FUN, frequency = daily, cache.dir = "cache",
  cache.name = "Cache", envir = parent.frame(), wait = FALSE, ...)
}
\arguments{
\item{FUN}{the function used to laod the data.}

\item{frequency}{how frequently should the cache expire.}

\item{cache.dir}{the directory containing the cached data files.}

\item{envir}{the enviroment into which data will be loaded.}

\item{wait}{should the function wait until stale data is refreshed.}

\item{...}{other parameters passed to \code{FUN}.}
}
\description{
This function maintains a data cache. When called, it will see if a "fresh"
data file is available. If not (i.e. the cache is "stale"), \code{FUN} will
be called to retreive more up-to-date data. On systems supporting forking
(e.g. Linux and Mac OS X), the refreshing of data will occur in the background
and the most "fresh" data will be returned until that background process
completes.
}
\details{
The \code{FUN} parameter is required. This function should return a named
\code{\link{list}} of objects. These objects will be assigned to the
given environment (the calling environment by default).

There are a number of frequencies available to determine when a cached file
becomes stale including: \code{\link{hourly}}, \code{\link{daily}},
\code{\link{weekly}}, \code{\link{monthly}}, \code{\link{yearly}},
\code{\link{nMinutes}}, \code{\link{nHours}}, and \code{\link{nDays}}.

It is possible to define custom frequencies. Each frequency function takes
one parameter, \code{timestamp}, which has a class type of \code{POSIXct},
and returns a \code{logical} value where \code{TRUE} indicates the cache
created at \code{timestamp} is stale.

See the Vignette (\code{vignette('DataCache')}) or demo
(\code{demo('WeatherCache')}) for more information.
}
\examples{
\dontrun{
library('weatherData')
loadWeatherData <- function(station_id='ALB') {
		results <- list(getDetailedWeather(station_id, Sys.Date()))
		names(results) <- paste0('weather.', station_id)
		return(results)
}
data.cache(loadWeatherData)
head(weather.ALB)
}
}
\seealso{
\link{daily}, \link{hourly}, \link{weekly}, \link{monthly},
  \link{yearly}, \link{nMinutes}, \link{nHours}, \link{nDays}
}

