% Generated by roxygen2 (4.0.1): do not edit by hand
\name{monthly}
\alias{monthly}
\title{Refresh data monthly.}
\usage{
monthly(timestamp)
}
\arguments{
\item{timestamp}{the timestamp to test whether the cache is stale.}
}
\description{
Refresh data monthly.
}
\seealso{
Other frequencies: \code{\link{daily}};
  \code{\link{hourly}}; \code{\link{nDays}};
  \code{\link{nHours}}; \code{\link{nMinutes}};
  \code{\link{now}}; \code{\link{weekly}};
  \code{\link{yearly}}
}

