% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies.R
\name{nDays}
\alias{nDays}
\title{Refresh every n days.}
\usage{
nDays(nDays)
}
\arguments{
\item{nDays}{number of days (minimally) between updates.}
}
\description{
Data becomes stale after n days.
}
\seealso{
Other frequencies: \code{\link{daily}},
  \code{\link{hourly}}, \code{\link{monthly}},
  \code{\link{nHours}}, \code{\link{nMinutes}},
  \code{\link{now}}, \code{\link{weekly}},
  \code{\link{yearly}}
}

