% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequencies.R
\name{nHours}
\alias{nHours}
\title{Refresh every n hours.}
\usage{
nHours(nHours)
}
\arguments{
\item{nHours}{number of hours (minimally) between updates.}
}
\description{
Data becomes stale after n hours.
}
\seealso{
Other frequencies: \code{\link{daily}},
  \code{\link{hourly}}, \code{\link{monthly}},
  \code{\link{nDays}}, \code{\link{nMinutes}},
  \code{\link{now}}, \code{\link{weekly}},
  \code{\link{yearly}}
}

