% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nMinutes}
\alias{nMinutes}
\title{Refresh every n minutes.}
\usage{
nMinutes(nMinutes)
}
\arguments{
\item{nMinutes}{number of minutes (minimally) between updates.}
}
\description{
Refresh every n minutes.
}
\seealso{
Other frequencies: \code{\link{daily}};
  \code{\link{hourly}}; \code{\link{monthly}};
  \code{\link{nDays}}; \code{\link{nHours}};
  \code{\link{now}}; \code{\link{weekly}};
  \code{\link{yearly}}
}

